pyproject.toml
[tool.poetry]
name = "langchain-super-basics"
version = "0.1.0"
description = ""
authors = ["Yuki Oshima <39944763+os1ma@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.338"
openai = "^1.3.3"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.1"
duckduckgo-search = "^3.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
